204: Try some time optimizations
Tried adding logic to run the sieve on only on odd numbers, since the
even numbers will already be filtered out during 2's iteration.
Time not decreased, it increased by ~ 0.0003 s after adding the logic.
Now as I type this out, I find this logic to be redundant, since all the
even numbers are already crossed out during 2's iteration and the
already present logic only runs the sieve if the number is not
filtered.
Damn.